My Project 1.1.0
This is the new feature release of My Project.

Here I can write a full description of what happened, what this feature
is all about etc etc. How does this works with "--add-description"?

The first line of commit, "My Project 1.1.0" will be used as the release
title. The third line onwards is the release description.